chore: change file name
CreateClass -> ClassCreate
